dom/webauthn/tests/browser/frame_webauthn_success.html

<!DOCTYPE html>
<meta charset=utf-8>
<head>
  <title>Full-run test for MakeCredential/GetAssertion for W3C Web Authentication</title>
  <script type="text/javascript" src="u2futil.js"></script>
  <script type="text/javascript" src="../pkijs/common.js"></script>
  <script type="text/javascript" src="../pkijs/asn1.js"></script>
  <script type="text/javascript" src="../pkijs/x509_schema.js"></script>
  <script type="text/javascript" src="../pkijs/x509_simpl.js"></script>
  <script type="text/javascript" src="../cbor/cbor.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>

<h1>Full-run test for MakeCredential/GetAssertion for W3C Web Authentication</h1>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1265472">Mozilla Bug 1265472</a>

<script class="testbody" type="text/javascript">
"use strict";

let gCredentialChallenge = new Uint8Array(16);
window.crypto.getRandomValues(gCredentialChallenge);
let gAssertionChallenge = new Uint8Array(16);
window.crypto.getRandomValues(gAssertionChallenge);

// the WebAuthn browser chrome tests watch for an element 'result' to appear in the DOM
function signalCompletion(aText) {
  console.log(aText)
  let result = document.createElement('h1');
  result.id = "result";
  result.textContent = aText;
  document.body.append(result);
}

function util_decodeCreatedCredential(aCredInfo) {
  /* PublicKeyCredential : Credential
     - rawId: Key Handle buffer pulled from U2F Register() Response
     - response : AuthenticatorAttestationResponse : AuthenticatorResponse
       - attestationObject: CBOR object
       - clientDataJSON: serialized JSON
     - clientExtensionResults: (not yet supported)
  */

  return webAuthnDecodeAttestation(aCredInfo.response.attestationObject.buffer)
  .then(function(decodedResult) {
    aCredInfo.publicKeyHandle = decodedResult.publicKeyHandle;
    aCredInfo.attestationObject = decodedResult.attestationObject;
    return aCredInfo;
  });
}

function util_checkAssertionAndSigValid(aPublicKey, aAssertion) {
  /* PublicKeyCredential : Credential
     - rawId: ID of Credential from AllowList that succeeded
     - response : AuthenticatorAssertionResponse : AuthenticatorResponse
       - clientDataJSON: serialized JSON
       - authenticatorData: RP ID Hash || U2F Sign() Response
       - signature: U2F Sign() Response
  */

  // Parse the signature data
  if (aAssertion.response.signature[0] != 0x01) {
    throw "User presence byte not set";
  }

  let presenceAndCounter = aAssertion.response.signature.slice(0,5);
  let signatureValue = aAssertion.response.signature.slice(5);

  let rpIdHash = aAssertion.response.authenticatorData.slice(0,32);

  // Assemble the signed data and verify the signature
  return deriveAppAndChallengeParam(window.location.host, aAssertion.response.clientDataJSON)
  .then(function(aParams) {
    console.log(aParams.appParam, rpIdHash, presenceAndCounter, aParams.challengeParam);
    console.log("ClientData buffer: ", hexEncode(aAssertion.response.clientDataJSON));
    console.log("ClientDataHash: ", hexEncode(aParams.challengeParam));
    return assembleSignedData(aParams.appParam, presenceAndCounter, aParams.challengeParam);
  })
  .then(function(aSignedData) {
    console.log(aPublicKey, aSignedData, signatureValue);
    return verifySignature(aPublicKey, aSignedData, signatureValue);
  })
}

let rp = {id: document.domain, name: "none", icon: "none"};
let user = {id: "none", name: "none", icon: "none", displayName: "none"};
let param = {type: "public-key", algorithm: "ES256"};
let makeCredentialOptions = {
  rp: rp,
  user: user,
  challenge: gCredentialChallenge,
  parameters: [param]
};

let credm = navigator.credentials;
credm.create({publicKey: makeCredentialOptions})
.then(util_decodeCreatedCredential)
.then(function testAssertion(aCredInfo) {
  let newCredential = {
    type: "public-key",
    id: Uint8Array.from(aCredInfo.rawId),
    transports: ["usb"],
  }

  let publicKeyCredentialRequestOptions = {
    challenge: gAssertionChallenge,
    timeout: 5000, // the minimum timeout is actually 15 seconds
    rpId: document.domain,
    allowList: [newCredential]
  };

  return credm.get({publicKey: publicKeyCredentialRequestOptions})
  .then(function(aAssertion) {
    /* Pass along the pubKey. */
    return util_checkAssertionAndSigValid(aCredInfo.publicKeyHandle, aAssertion);
  });
})
.then(function(aSigVerifyResult) {
  signalCompletion("Signing signature verified: " + aSigVerifyResult);
})
.catch(function failure(aReason) {
  signalCompletion("Failure: " + aReason);
});

</script>

</body>
</html>
